pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dtu-findit-webscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.189.0",
    "google-cloud-storage>=2.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pyarrow>=18.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "pymupdf>=1.25.0",
    "pytesseract>=0.3.10",
    "rapidfuzz>=3.0.0",
    "requests>=2.32.5",
    "selenium>=4.40.0",
    "lxml>=5.0.0",
    "tqdm>=4.0.0",
    "ipykernel>=7.2.0",
]

[tool.setuptools.packages.find]
where = ["src"]
